StreamLoader Monitoring : Add Latency to Stdout
The metrics was added to StreamLoader in #165. The `Latency` metric however was only added to `Statsd` and not `stdout`. This PR adds the missing latency to `stdout`. This is particularly important for GCE hosts running on Container optimised OSs and local testsing. COS images do not support GCP ops agent which is prerequisite for Statsd plugin.