Improve exception handling immediately after altering the table
When the BigQuery Loader encounters a new schema for the first time, it alters the table to add new
columns.  But there can be a delay between adding the new column, and the column becoming available
for receiving new events.

Opening a Writer is not considered successful, until it is opened in a state where it knows about
the newly added column.  This commit adds a delay/backoff when opening the Writer, so it retries
gracefully.