Recover from server-side schema mismatch exceptions
This is about what happens immediately after altering the table to add a
new column.

In #366 I amended the loader so it retries opening the Writer, until the
Writer discovers the extra columns.  However, this only accounted for
the **client-side** writer discovering the columns.

When running the loader, I discovered there is also an issue where the
**server-side** end of the connection can be slow to discover the newly
added columns.  This commit catches and retries the known exception when
the server has not yet discovered the new columns.