Allow for delay in Writer discovering new columns
After the loader alters the table to add new columns, it immediately
opens a new Writer and expects the Writer to be aware of the new
columns. However, we have found the Writer might get opened with no
awareness of the newly added columns. Presumbably because of the async
nature of BigQuery's architecture.

This fix works by retrying opening the writer until eventually it should
get opened with awareness of the new columns