Feature flag to support the legacy column style
In v1 of BigQuery Loader, we created a separate column for each _minor_
version of an Iglu schema.  Whereas in v2 of BigQuery Loader, schemas
with the same major version share the same column.  This might make it
hard for users to upgrade from version 1 to version 2; and so users
might miss out on all the other lovely stuff we added in v2.

This PR adds back in support for the legacy style of columns.  It is
enabled by setting `legacyColumns = true` in the config file.