Fix startup on missing existing table
During startup, the loader tries to fetch details of an existing table.
It only creates a new table if the previous table does not exist.

But this was broken because we were not properly checking the response
returned from BigQuery when fetching table details.  If a table does not
exist then the client returns a null, but does not throw an exception.