Loader should handle BigQuery exceptions for an invalid request (#392)
* Loader should handle BigQuery exceptions for an invalid request

BigQuery occasionally returns a response with a 400 response code. Before
this PR, it would cause the loader to crash and exit. After this PR, we
break the original request into smaller requests and try to load each
event individually. This allows valid events in the batch to get loaded,
while the invalid requests eventually become failed inserts.

Known examples of when this needed: When the table is partitioned by
`derived_tstamp`, and an event has a `derived_tstamp` more than 1825
days in the past.

* Install sbt in github actions