Fix and improve ProcessingSpec for legacy column mode
A few of the tests in `ProcessingSpec` had recently been disabled by
replacing the real test with `TestControl.executeEmbed(io.timeout(10.seconds))`.
This PR fixes and re-enables the test, so we have better coverage of
when legacy modes are enabled.

Also, I make more use of the new ability to test `writtenToBQ` which was
added in #394. This means we have more test coverage that the expected
fields get loaded.