Open several writers to increase BigQuery throughput
Before this PR, the loader opened a single `JsonStreamWriter`, and
therefore a single stream (single bidirectional RPC). We observed the
loader seems to hit a bottleneck where it cannot write into BigQuery
quickly enough.

After this PR, the loader opens a configurable number of streams, which
is expected to prevent the bottleneck.

This implementation closely follows a similar change we made in the
snowflake loader: snowplow-incubator/snowflake-loader#57