Add parallelism to parseBytes and transform (#400)
For pods with more than 1 CPU, tests have shown that we get a better throughput
when adding some parallelism on CPU-intensive steps.

Actually, tests have shown that we get the best throughput
when adding parallelism to transform step only.

I wonder if we should remove the parallelism on parseBytes,
but given that it is there now, I'd be in favor of keeping it with a low default,
so that there is no parallelism, and if 1 day we want to change it we can.

Also, we've observed better throughput with writeBatchConcurrency = 2,
so I updated the default.